Make some refactor & lines length <= 90

+++ GetSunsetRise.cpp
@@ -14,7 +14,9 @@ GetSunsetRise::GetSunsetRise(QWidget *pwgt)
     setupUi(this);
     QFile file(QString(":/timeZonesRC.txt"));
     if(!file.open(QIODevice::ReadOnly)) {
-        QMessageBox::critical(nullptr, tr("ERROR"), tr("Cannot open file ") + file.fileName());
+        QMessageBox::critical(nullptr,
+                              tr("ERROR"),
+                              tr("Cannot open file ") + file.fileName());
     }
     QTextStream in(&file);
     while(!in.atEnd()) {
@@ -25,7 +27,9 @@ GetSunsetRise::GetSunsetRise(QWidget *pwgt)
     QDir dir;
     if(!QDir("scripts").exists()) {
         if(!dir.mkdir("scripts")) {
-            QMessageBox::critical(nullptr, tr("ERROR"), tr("Cannot create directory \"scripts\""));
+            QMessageBox::critical(nullptr,
+                                  tr("ERROR"),
+                                  tr("Cannot create directory \"scripts\""));
         }
     }
 
@@ -54,17 +58,23 @@ void GetSunsetRise::calc()
     m_sunSetRise->setEastOrWest(rbWest->isChecked());
 
     if(rbSunSet->isChecked()) {
-        lResults->setText(tr("Sunset at ") + toHumanTime(static_cast<double>(m_sunSetRise->getConfiguredTime())));
+        lResults->setText(
+                    tr("Sunset at ") + toHumanTime(m_sunSetRise->getConfiguredTime()));
     } else {
-        lResults->setText(tr("Sunrise at ") + toHumanTime(static_cast<double>(m_sunSetRise->getConfiguredTime())));
+        lResults->setText(
+                    tr("Sunrise at ") + toHumanTime(m_sunSetRise->getConfiguredTime()));
     }
 }
 
 void GetSunsetRise::clickedCity(QAction *pAction)
 {
-    QFile file(QString("scripts/") + QString(pAction->property("cityName").toString()) + QString(".js"));
+    QFile file(QString("scripts/") +
+               QString(pAction->property("cityName").toString()) +
+               QString(".js"));
     if(!file.open(QIODevice::ReadOnly)) {
-        QMessageBox::critical(nullptr, tr("ERROR"), tr("Cannot open file ") + file.fileName());
+        QMessageBox::critical(nullptr,
+                              tr("ERROR"),
+                              tr("Cannot open file ") + file.fileName());
 
         return;
     }
@@ -106,7 +116,9 @@ void GetSunsetRise::clickedCity(QAction *pAction)
 qreal GetSunsetRise::zoneToTime(const int &index)
 {
     if(m_timeZones.size() < index + 1) {
-        QMessageBox::critical(nullptr, tr("ERROR"), tr("The time zone was selected an incorrect way!"));
+        QMessageBox::critical(nullptr,
+                              tr("ERROR"),
+                              tr("The time zone was selected an incorrect way!"));
 
         return 0.0;
     }
@@ -117,7 +129,10 @@ qreal GetSunsetRise::zoneToTime(const int &index)
 QString GetSunsetRise::toHumanTime(double preResult)
 {
     QString hours = QString::number(static_cast<int>(preResult));
-    QString minutes = QString::number(preResult - static_cast<double>(static_cast<int>(preResult)), 'f', 2);
+    QString minutes =
+            QString::number(preResult - static_cast<double>(static_cast<int>(preResult)),
+                            'f',
+                            2);
 
     switch (hours.size()) {
     case 0:
